Scripting: Start work on integrating QuickJS for scripting functionality
Scripts can already be run, but they aren't integrated with the mqtt
system yet. The idea is to allow the user to add scripts via a REST API
that are then run based on time-based or mqtt-based events. There should
even be an option to run the scripts in a synchronous manner which
allows them to interact with the packets, e.g. stopping a published
message before it gets send to the clients.

In the future, we want to support multiple scripting languages/engines.
QuickJS was chosen as a first language because it's relatively easy to
integrate and - most importantly - supports timeouts/interrupting the
scripts. This is very important, otherwise a bad script could hang the
whole broker.