Fix publishing client being disconnected if sending fails
This means that previously when we failed to send some data, e.g.
because the client is already disconnected, we disconnected the
publishing client because the exception would just travel up the call
stack.

This was fixed by catching the execeptions in
ClientThreadManager::sendData and notifying the Application that the
client disconnected. Handling disconnectes synchronously wasn't possible
anymore in this system as it might lead to deadlocks, so we just set a
bool in the client now and after every publish check every client if
there was an error.

This might cause issues when scaling to high amounts of clients, but we
can still improve this design then - even with the current system,
iterating over an array with like 1000 to 10000 elements is still pretty
quick and more clients make problems anyways.