WIP Refactor Application to ApplicationState, simplifying lock patterns
We now have one ApplicationState which manages all states. If someone
wants to modify its state, he needs to call
ApplicationState::requestChange, which stores the request in a lockless
queue. The ApplicationState thread then handles all requests one by one.

This effectively merges Application and MQTTPersistentState while also
optimizing a lot of inefficient things. For example, it's now possible
to store a pointer to the MQTTClientState withing the epoll event
struct. Pointer validity is always ensured, because the only way to
delete such a pointer is to suspend all threads and then resume them.

This new implementation is still missing a lot of features, most notably
scripts and timeouts.