Publish broker log on $NIOEV/log
This allows a remote client to see the broker's log. This might be
security critical, so it might be smart to add a config to restrict
specific users to this topic (or all $NIOEV/# topics).

To avoid endless loops in debug mode, information about published
messages to $NIOEV/log isn't logged itself. Maybe we should disallow
clients publishing to this topic in the first place?

This patch is the first step towards allowing remote script development.
Next, we need to allow subscribing with a websocket api to mqtt topics
to view the log.