Dynamically scale sleep duration in worker thread
Because we want to avoid the syscall, we can't use a condition varaible
to wait until new data is the queue. Instead, we used to always wait
1ms. This was an okay-ish compromise, but does leave some performance on
the table when high throughput is required (as we can in theory wait for
1ms until a slot in the queue becomes free) and spins the CPU needlessly
when there is nothing to do. The latter case is only a concern for
low-powered devices; I have an old atom CPU were this spinning caused
about 2-3% CPU usage on one core.

Now, we dynamically scale our sleep duration. We yield 5 times, then we
sleep 200 times for 10µs then sleep 100 times for 1ms. After that, we
sleep for 10ms continously. When we receive data, we step one level down
again (10ms -> 1ms -> 10µs -> yield) and reset the counter for how many
times we slept at the current interval (used for increasing the level).
This means that, at worst, we could wait 10ms to find a slot, but after
that the worker thread spins slowly up and if the load stays high, he
will stay active for longer than before. Another strategy where we
immediately jump to yield instead of slowly scaling would also be an
option, maybe we can make this configurable?

We yield for 5 times because my testing showed that after yielding 5
times, we almost always jump to the microseconds-category next, even if
we yield 100 times. We also need to keep in mind that yielding is
actually really bad for the CPU, as it needlessly steals CPU time from
other processes (and our own threads) if we yield for too long.

I don't know if these results transfer to other CPUs/systems, maybe
these steps should be configurable as well?