Split up packets into components internally & fix QoS2 bugs
This is a quite a big patch that splits up the SendTask struct into a
new EncodedPacket system. The idea is that when publishing the same
message to different clients with different packet identifiers & ids, we
can reuse the topic buffer and the propertylist+payload buffer. To send
the data, we use sendmsg, which is basically a scatter send, allowing us
to specify a range of buffers to send from. Using the sent bytes value
we can then piece together what was send.

This should allow for greater performance and also makes the code
cleaner in some parts. For example, when sending data that has been
queued up, we now just do a single scatter send and don't need to loop
over individual buffers anymore.

This patch also fixes 2 bugs that I discovered by accident. For once,
when a client reconnected with a session present, we used to sent
256*256 PUBCOMP packets for every client id. The other bug was a bit of
improper spec implementation in QoS2 handling, when receiving a message
with QoS2 that we already did receive we used to publish that thing
twice - instead of dropping it as the spec requires.

The big missing thing now is sane packet id generation/assignment, then
we can move on to more MQTTv5 features.